Change signature of get_max()

+++ lib/algorithm/algorithm.cpp
@@ -21,7 +21,7 @@ bool approx_equal(const double a, const double b)
     return fabs(a - b) <= delta;
 }
 
-long long get_max(long long a, long long b)
+int get_max(int a, int b)
 {
     if (a > b) return a;
     return b;
